// src/helpers/types.ts
export enum BaixariaPallete {
    // Background e textos gerais
    Header = "#101010",
    Footer = "#101010",
    Background = "#141414", 
  
    PrimaryText = "#DCDDD4",
    SecondaryText = "#DCDDD4", // tom intermediário para textos secundários
    DisabledText = "rgba(100, 142, 135, 0.6)", // versão suavizada do secundário
    IconDefault = '#DAF1DE',
    ErrorText = "#FF6B6B",
  
    // Caixas e containers
    DashboardBox = "#060F10", // mantendo a uniformidade com o background
    Box = 'rgba(255, 255, 255, 0.1)', // usando a cor de acento com opacidade para reduzir o contraste
    Border = "#EBE3FF",
  
    // Botões
    PrimaryButton = "#89FF8E",
    SecondaryButton = "#A7E2A9",
    WarningButton = "#F39C12",
    AlertButton = "#FDFF89",
    SuccessButton = "#25714D",
    InfoButton = '#3483f4',
    ErrorButton = "#E74C3C",
    ButtonHover = "#89FF8E",
  
    PrimaryButtonHover = "#6ECC71",
    SecondaryButtonHover = "#86B587",
    WarningButtonHover = "#C27D0E",
    AlertButtonHover = "#CACC6E",
    SuccessButtonHover = "#1E5A3E",
    InfoButtonHover = "#2A69C3",
    ErrorButtonHover = "#B93D30",
  
    // Modais e inputs
    ModalPopup = "#060F10",
    InputField = "#DCDDD4",
    WarningInputField = '#FFE9D5',
    ErrorInputField = '#ffebee',
    SuccessInputField = '#DAF1DE',
    PlaceholderText = "#CBE4DE",
  
    // Scrollbars e detalhes adicionais
    Scrollbar = "#648E87"
  }